#18058: Move init_sync_registers to TRISC0
Initializing all the sync registers takes around 120 cycles on wormhole. By
moving the work to TRISC0 (when it's already idle) we can cut the cost to BRISC
by around 100 cycles.
